+++ packages/round-manager/src/features/round/RoundDetailForm.tsx
@@ -1,99 +1,17 @@
-import { Fragment, useContext, useState, FC } from "react";
+import { useContext, useState } from "react";
 import { yupResolver } from "@hookform/resolvers/yup";
 import "react-datetime/css/react-datetime.css";
 import {
-  Control,
-  Controller,
-  FieldErrors,
   SubmitHandler,
-  useController,
   useForm,
-  UseFormRegisterReturn,
 } from "react-hook-form";
-import * as yup from "yup";
-import { Listbox, RadioGroup, Transition } from "@headlessui/react";
-import {
-  CheckIcon,
-  SelectorIcon,
-} from "@heroicons/react/solid";
-import { Input } from "common/src/styles";
-import moment from "moment";
 import { Program, Round } from "../api/types";
-import { SupportType } from "../api/utils";
 import { FormStepper } from "../common/FormStepper";
 import { FormContext } from "../common/FormWizard";
 import _ from "lodash";
-import { classNames } from "common";
-
-const ValidationSchema = yup.object().shape({
-  roundMetadata: yup.object({
-    name: yup
-      .string()
-      .required("This field is required.")
-      .min(8, "Round name must be at least 8 characters."),
-    roundType: yup.string().required("You must select the round type."),
-    support: yup.object({
-      type: yup
-        .string()
-        .required("You must select a support type.")
-        .notOneOf(
-          ["Select what type of input."],
-          "You must select a support type."
-        ),
-      info: yup
-        .string()
-        .required("This field is required.")
-        .when("type", {
-          is: "Email",
-          then: yup
-            .string()
-            .email()
-            .required("You must provide a valid email address."),
-        })
-        .when("type", {
-          is: (val: string) => val && val != "Email",
-          then: yup.string().url().required("You must provide a valid URL."),
-        }),
-    }),
-  }),
-  applicationsStartTime: yup
-    .date()
-    .required("This field is required.")
-    .min(new Date(), "You must enter a date and time in the future.")
-    .max(
-      yup.ref("applicationsEndTime"),
-      "Applications start date must be earlier than the applications end date"
-    ),
-  applicationsEndTime: yup
-    .date()
-    .required("This field is required.")
-    .min(
-      yup.ref("applicationsStartTime"),
-      "Applications end date must be later than applications start date"
-    )
-    .max(
-      yup.ref("roundStartTime"),
-      "Applications end date must be earlier than the round start date"
-    ),
-  roundStartTime: yup
-    .date()
-    .required("This field is required.")
-    .min(
-      yup.ref("applicationsEndTime"),
-      "Round start date must be later than applications end date"
-    )
-    .max(
-      yup.ref("roundEndTime"),
-      "Round start date must be earlier than the round end date"
-    ),
-  roundEndTime: yup
-    .date()
-    .required("This field is required.")
-    .min(
-      yup.ref("roundStartTime"),
-      "Round end date must be later than the round start date"
-    ),
-});
+import { ValidationSchema } from "./ApplicationFormValidationSchema";
+import { RoundName, ProgramChain, ContactInformation, BetterDatetime, RoundType, Support } from "./ApplicationFormComponents";
+import moment from "moment";
 
 interface RoundDetailFormProps {
   stepper: typeof FormStepper;
@@ -103,7 +21,7 @@ interface RoundDetailFormProps {
 export function RoundDetailForm(props: RoundDetailFormProps) {
   const program = props.initialData?.program;
   const { currentStep, setCurrentStep, stepsCount, formData, setFormData } =
-    useContext(FormContext);
+  useContext(FormContext);
   const defaultRoundMetadata = {
     ...((formData as Partial<Round>)?.roundMetadata ?? {}),
     feesPercentage: 0,
@@ -123,10 +41,9 @@ export function RoundDetailForm(props: RoundDetailFormProps) {
   });
 
   const FormStepper = props.stepper;
-  const [applicationStartDate, setApplicationStartDate] = useState<any>();
-  const [applicationEndDate, setApplicationEndDate] = useState<any>();
-  const [roundStartDate, setRoundStartDate] = useState<any>();
-  const [roundEndDate, setRoundEndDate] = useState();
+  const [applicationStartDate, setApplicationStartDate] = useState<moment.Moment>();
+  const [applicationEndDate, setApplicationEndDate] = useState<moment.Moment>();
+  const [roundStartDate, setRoundStartDate] = useState<moment.Moment>();
 
   const next: SubmitHandler<Round> = async (values) => {
     const data = _.merge(formData, values);
@@ -134,30 +51,9 @@ export function RoundDetailForm(props: RoundDetailFormProps) {
     setCurrentStep(currentStep + 1);
   };
 
-  const now = moment().add(1, "hour").startOf("hour");
   const prev = () => setCurrentStep(currentStep - 1);
-  const yesterday = moment().subtract(1, "day");
 
-  const disablePastDate = (current: moment.Moment) => {
-    return current.isAfter(yesterday);
-  };
-
-  function disableBeforeApplicationStartDate(current: moment.Moment) {
-    return current.isAfter(applicationStartDate);
-  }
-
-  const disablePastAndBeforeRoundStartDate = (current: moment.Moment) => {
-    return disablePastDate(current);
-  };
-
-  function disableBeforeApplicationEndDate(current: moment.Moment) {
-    return current.isAfter(applicationEndDate);
-  }
-
-  function disableBeforeRoundStartDate(current: moment.Moment) {
-    return current.isAfter(roundStartDate);
-  }
-
+  const now = moment(); 
 
   return (
     <div>
@@ -197,7 +93,7 @@ export function RoundDetailForm(props: RoundDetailFormProps) {
                     control={control}
                   />
                 </div>
-                <div className="col-span-6 sm:col-span-3">
+                <div className="col-span-6 sm:col-span-3 pt-2">
                   <ContactInformation
                     register={register("roundMetadata.support.info")}
                     errors={errors}
@@ -222,10 +118,10 @@ export function RoundDetailForm(props: RoundDetailFormProps) {
                 <div className="col-span-6 sm:col-span-3">
                   <div
                     className={`relative border rounded-md px-3 py-2 mb-2 shadow-sm focus-within:ring-1 ${
-                      errors.applicationsStartTime
-                        ? "border-red-300 text-red-900 placeholder-red-300 focus-within:outline-none focus-within:border-red-500 focus-within: ring-red-500"
-                        : "border-gray-300 focus-within:border-indigo-600 focus-within:ring-indigo-600"
-                    }`}
+errors.applicationsStartTime
+? "border-red-300 text-red-900 placeholder-red-300 focus-within:outline-none focus-within:border-red-500 focus-within: ring-red-500"
+: "border-gray-300 focus-within:border-indigo-600 focus-within:ring-indigo-600"
+}`}
                   >
                     <BetterDatetime
                       control={control}
@@ -233,6 +129,7 @@ export function RoundDetailForm(props: RoundDetailFormProps) {
                       label="Start Date"
                       date={applicationStartDate}
                       setDate={setApplicationStartDate}
+                      minDate={now}
                     />
                   </div>
                   {errors.applicationsStartTime && (
@@ -248,10 +145,10 @@ export function RoundDetailForm(props: RoundDetailFormProps) {
                 <div className="col-span-6 sm:col-span-3">
                   <div
                     className={`relative border rounded-md px-3 py-2 mb-2 shadow-sm focus-within:ring-1 ${
-                      errors.applicationsEndTime
-                        ? "border-red-300 text-red-900 placeholder-red-300 focus-within:outline-none focus-within:border-red-500 focus-within: ring-red-500"
-                        : "border-gray-300 focus-within:border-indigo-600 focus-within:ring-indigo-600"
-                    }`}
+errors.applicationsEndTime
+? "border-red-300 text-red-900 placeholder-red-300 focus-within:outline-none focus-within:border-red-500 focus-within: ring-red-500"
+: "border-gray-300 focus-within:border-indigo-600 focus-within:ring-indigo-600"
+}`}
                   >
                     <BetterDatetime 
                       control={control}
@@ -259,6 +156,7 @@ export function RoundDetailForm(props: RoundDetailFormProps) {
                       label="End Date"
                       date={applicationEndDate}
                       setDate={setApplicationEndDate}
+                      minDate={applicationStartDate}
                     />
                   </div>
                   {errors.applicationsEndTime && (
@@ -281,10 +179,10 @@ export function RoundDetailForm(props: RoundDetailFormProps) {
                 <div className="col-span-6 sm:col-span-3">
                   <div
                     className={`relative border rounded-md px-3 py-2 mb-2 shadow-sm focus-within:ring-1 ${
-                      errors.roundStartTime
-                        ? "border-red-300 text-red-900 placeholder-red-300 focus-within:outline-none focus-within:border-red-500 focus-within: ring-red-500"
-                        : "border-gray-300 focus-within:border-indigo-600 focus-within:ring-indigo-600"
-                    }`}
+errors.roundStartTime
+? "border-red-300 text-red-900 placeholder-red-300 focus-within:outline-none focus-within:border-red-500 focus-within: ring-red-500"
+: "border-gray-300 focus-within:border-indigo-600 focus-within:ring-indigo-600"
+}`}
                   >
                     <BetterDatetime
                       control={control}
@@ -292,6 +190,7 @@ export function RoundDetailForm(props: RoundDetailFormProps) {
                       label="Start Date"
                       date={roundStartDate}
                       setDate={setRoundStartDate}
+                      minDate={applicationEndDate}
                     />
                   </div>
                   {errors.roundStartTime && (
@@ -307,15 +206,16 @@ export function RoundDetailForm(props: RoundDetailFormProps) {
                 <div className="col-span-6 sm:col-span-3">
                   <div
                     className={`relative border rounded-md px-3 py-2 mb-2 shadow-sm focus-within:ring-1 ${
-                      errors.roundEndTime
-                        ? "border-red-300 text-red-900 placeholder-red-300 focus-within:outline-none focus-within:border-red-500 focus-within: ring-red-500"
-                        : "border-gray-300 focus-within:border-indigo-600 focus-within:ring-indigo-600"
-                    }`}
+errors.roundEndTime
+? "border-red-300 text-red-900 placeholder-red-300 focus-within:outline-none focus-within:border-red-500 focus-within: ring-red-500"
+: "border-gray-300 focus-within:border-indigo-600 focus-within:ring-indigo-600"
+}`}
                   >
                     <BetterDatetime 
                       control={control}
                       name="roundEndTime"
                       label="End Date"
+                      minDate={roundStartDate}
                     />
                   </div>
                   {errors.roundEndTime && (
@@ -330,7 +230,6 @@ export function RoundDetailForm(props: RoundDetailFormProps) {
               </div>
             </div>
 
-            {/* Round Type */}
             <div className="p-6 bg-white">
               <div className="grid grid-rows-1">
                 <p>
@@ -368,414 +267,3 @@ export function RoundDetailForm(props: RoundDetailFormProps) {
   );
 }
 
-function RoundName(props: {
-  register: UseFormRegisterReturn<string>;
-  errors: FieldErrors<Round>;
-}) {
-  return (
-    <div className="col-span-6 sm:col-span-3">
-      <div className="flex justify-between">
-        <label htmlFor="roundMetadata.name" className="text-sm">
-          Round Name
-        </label>
-        <span className="text-right text-violet-400 float-right text-xs mt-1">
-          *Required
-        </span>
-      </div>
-      <Input
-        {...props.register}
-        className={"h-10"}
-        $hasError={props.errors.roundMetadata?.name}
-        type="text"
-        id={"roundMetadata.name"}
-      />
-      {props.errors.roundMetadata?.name && (
-        <p className="text-xs text-pink-500">
-          {props.errors.roundMetadata?.name?.message}
-        </p>
-      )}
-    </div>
-  );
-}
-
-function ProgramChain(props: { program: Program }) {
-  const { program } = props;
-  return (
-    <div className="col-span-6 sm:col-span-3 opacity-50">
-      <Listbox disabled>
-        <div>
-          <Listbox.Label className="block text-sm">Program Chain</Listbox.Label>
-          <div className="relative mt-1">
-            <Listbox.Button
-              className={`relative w-full cursor-default rounded-md border border-gray-300 bg-white py-2 pl-3 pr-10 text-left shadow-sm sm:text-sm h-10`}
-            >
-              <span className="flex items-center">
-                {program.chain?.logo && (
-                  <img
-                    src={program.chain.logo}
-                    alt="chain logo"
-                    data-testid="chain-logo"
-                    className="h-5 w-5 flex-shrink-0 rounded-full"
-                  />
-                )}
-                {
-                  <span className="ml-3 block truncate">
-                    {program.chain?.name}
-                  </span>
-                }
-              </span>
-            </Listbox.Button>
-          </div>
-        </div>
-      </Listbox>
-    </div>
-  );
-}
-
-function ContactInformation(props: {
-  register: UseFormRegisterReturn<string>;
-  errors: FieldErrors<Round>;
-}) {
-  return (
-    <div className="mt-2 mb-2">
-      <div className="flex justify-between">
-        <label htmlFor="roundMetadata.support.info" className="text-sm">
-          Contact Information
-        </label>
-        <span className="text-right text-violet-400 float-right text-xs mt-1">
-          *Required
-        </span>
-      </div>
-      <Input
-        {...props.register}
-        className={"h-10"}
-        $hasError={props.errors.roundMetadata?.support?.info}
-        type="text"
-        placeholder="Enter desired form of contact here. Ex: website, email..."
-        id={"roundMetadata.support.info"}
-      />
-      {props.errors.roundMetadata?.support?.info && (
-        <p className="text-xs text-pink-500">
-          {props.errors.roundMetadata?.support.info?.message}
-        </p>
-      )}
-    </div>
-  );
-}
-
-function SupportTypeButton(props: {
-  errors: FieldErrors<Round>;
-  supportType?: SupportType;
-}) {
-  const { supportType } = props;
-  return (
-    <Listbox.Button
-      className={`relative w-full cursor-default rounded-md border h-10 bg-white py-2 pl-3 pr-10 text-left shadow-sm ${
-        props.errors.roundMetadata?.support?.type
-          ? "border-red-300 text-red-900 placeholder-red-300 focus-within:outline-none focus-within:border-red-500 focus-within: ring-red-500"
-          : "border-gray-300 focus:border-indigo-500 focus:outline-none focus:ring-1 focus:ring-indigo-500 sm:text-sm"
-      }`}
-      data-testid="support-type-select"
-      id={"roundMetadata.support.type"}
-    >
-      <span className="flex items-center">
-        {supportType?.default ? (
-          <span className="ml-3 block truncate text-gray-400">
-            {supportType?.name}
-          </span>
-        ) : (
-          <span className="ml-3 block truncate">{supportType?.name}</span>
-        )}
-      </span>
-      <span className="pointer-events-none absolute inset-y-0 right-0 ml-3 flex items-center pr-2">
-        <SelectorIcon className="h-5 w-5 text-gray-400" aria-hidden="true" />
-      </span>
-    </Listbox.Button>
-  );
-}
-
-function SupportTypeDropdown(props: {
-  register: UseFormRegisterReturn<string>;
-  errors: FieldErrors<Round>;
-  control: Control<Round>;
-  supportTypes: SupportType[];
-}) {
-  const { field } = useController({
-    name: "roundMetadata.support.type",
-    defaultValue: props.supportTypes[0].name,
-    control: props.control,
-    rules: {
-      required: true,
-    },
-  });
-  return (
-    <div className="col-span-6 sm:col-span-3 relative mt-2">
-      <Listbox {...field}>
-        {({ open }) => (
-          <div>
-            <Listbox.Label className="text-sm mt-4 mb-2">
-              <p className="text-sm">
-                <span>Support Input</span>
-                <span className="text-right text-violet-400 float-right text-xs mt-1">
-                  *Required
-                </span>
-              </p>
-            </Listbox.Label>
-            <div className="mt-1 mb-2 shadow-sm block rounded-md border-gray-300 focus:border-indigo-500 focus:ring-indigo-500 sm:text-sm">
-              <SupportTypeButton
-                errors={props.errors}
-                supportType={props.supportTypes.find(
-                  (supportType) => supportType.name === field.value
-                )}
-              />
-              <Transition
-                show={open}
-                as={Fragment}
-                leave="transition ease-in duration-100"
-                leaveFrom="opacity-100"
-                leaveTo="opacity-0"
-              >
-                <Listbox.Options className="absolute z-10 mt-1 max-h-56 w-full overflow-auto rounded-md bg-white py-1 text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none sm:text-sm">
-                  {props.supportTypes.map(
-                    (type) =>
-                      !type.default && (
-                        <Listbox.Option
-                          key={type.name}
-                          className={({ active }) =>
-                            classNames(
-                              active
-                                ? "text-white bg-indigo-600"
-                                : "text-gray-900",
-                              "relative cursor-default select-none py-2 pl-3 pr-9"
-                            )
-                          }
-                          value={type.name}
-                          data-testid="support-type-option"
-                        >
-                          {({ selected, active }) => (
-                            <>
-                              <div className="flex items-center">
-                                <span
-                                  className={classNames(
-                                    selected ? "font-semibold" : "font-normal",
-                                    "ml-3 block truncate"
-                                  )}
-                                >
-                                  {type.name}
-                                </span>
-                              </div>
-
-                              {selected ? (
-                                <span
-                                  className={classNames(
-                                    active ? "text-white" : "text-indigo-600",
-                                    "absolute inset-y-0 right-0 flex items-center pr-4"
-                                  )}
-                                >
-                                  <CheckIcon
-                                    className="h-5 w-5"
-                                    aria-hidden="true"
-                                  />
-                                </span>
-                              ) : null}
-                            </>
-                          )}
-                        </Listbox.Option>
-                      )
-                  )}
-                </Listbox.Options>
-              </Transition>
-            </div>
-            {props.errors.roundMetadata?.support?.type && (
-              <p className="mt-2 text-xs text-pink-500">
-                {
-                  "You must select a support type."
-                  // TODO: Use YUP for error message
-                }
-              </p>
-            )}
-          </div>
-        )}
-      </Listbox>
-    </div>
-  );
-}
-
-function Support(props: {
-  register: UseFormRegisterReturn<string>;
-  errors: FieldErrors<Round>;
-  control: Control<Round>;
-}) {
-  // TODO: Add regex for URLs
-  const supportTypes: SupportType[] = [
-    {
-      name: "Select what type of input.",
-      regex: "https://www.google.com",
-      default: true,
-    },
-    {
-      name: "Email",
-      regex: "https://www.google.com",
-      default: false,
-    },
-    {
-      name: "Website",
-      regex: "https://www.google.com",
-      default: false,
-    },
-    {
-      name: "Discord Group Invite Link",
-      regex: "https://www.google.com",
-      default: false,
-    },
-    {
-      name: "Telegram Group Invite Link",
-      regex: "https://www.google.com",
-      default: false,
-    },
-    {
-      name: "Google Form Link",
-      regex: "https://www.google.com",
-      default: false,
-    },
-    {
-      name: "Other (please provide a link)",
-      regex: "https://www.google.com",
-      default: false,
-    },
-  ];
-
-  return (
-    <SupportTypeDropdown
-      register={props.register}
-      errors={props.errors}
-      control={props.control}
-      supportTypes={supportTypes}
-    />
-  );
-}
-
-function RoundType(props: {
-  register: UseFormRegisterReturn<string>;
-  control?: Control<Round>;
-}) {
-  const { field: roundTypeField } = useController({
-    name: "roundMetadata.roundType",
-    defaultValue: "",
-    control: props.control,
-    rules: {
-      required: true,
-    },
-  });
-
-  return (
-    <>
-      {" "}
-      <div className="col-span-6 sm:col-span-3">
-        <RadioGroup {...roundTypeField} data-testid="round-type-selection">
-          <div>
-            <RadioGroup.Option value="public" className="mb-2">
-              {({ checked, active }) => (
-                <span className="flex items-center text-sm">
-                  <span
-                    className={classNames(
-                      checked
-                        ? "bg-indigo-600 border-transparent"
-                        : "bg-white border-gray-300",
-                      active ? "ring-2 ring-offset-2 ring-indigo-500" : "",
-                      "h-4 w-4 rounded-full border flex items-center justify-center"
-                    )}
-                    aria-hidden="true"
-                  >
-                    <span className="rounded-full bg-white w-1.5 h-1.5" />
-                  </span>
-                  <RadioGroup.Label
-                    as="span"
-                    className="ml-3 block text-sm text-gray-700"
-                    data-testid="round-type-public"
-                  >
-                    Yes, make my round public
-                    <p className="text-xs text-gray-400">
-                      Anyone on the Gitcoin Explorer homepage will be able to
-                      see your round
-                    </p>
-                  </RadioGroup.Label>
-                </span>
-              )}
-            </RadioGroup.Option>
-            <RadioGroup.Option value="private">
-              {({ checked, active }) => (
-                <span className="flex items-center text-sm">
-                  <span
-                    className={classNames(
-                      checked
-                        ? "bg-indigo-600 border-transparent"
-                        : "bg-white border-gray-300",
-                      active ? "ring-2 ring-offset-2 ring-indigo-500" : "",
-                      "h-4 w-4 rounded-full border flex items-center justify-center"
-                    )}
-                    aria-hidden="true"
-                  >
-                    <span className="rounded-full bg-white w-1.5 h-1.5" />
-                  </span>
-                  <RadioGroup.Label
-                    as="span"
-                    className="ml-3 block text-sm text-gray-700"
-                    data-testid="round-type-private"
-                  >
-                    No, keep my round private
-                    <p className="text-xs text-gray-400">
-                      Only people with the round link can see your round.
-                    </p>
-                  </RadioGroup.Label>
-                </span>
-              )}
-            </RadioGroup.Option>
-          </div>
-        </RadioGroup>
-      </div>
-    </>
-  );
-}
-
-
-interface BetterDatetimeProps {
-  control: Control<Round>;
-  name: any;
-  label: string;
-  date?: moment.Moment;
-  setDate?: (date: moment.Moment) => void;  
-}
-
-const BetterDatetime: FC<BetterDatetimeProps> = ({ control, name, label, date, setDate }) => {
-  const now = moment().format('YYYY-MM-DDTHH:mm');
-  return (
-    <div className="relative w-full border-0 p-0 placeholder-grey-40 focus:ring-0 text-sm">
-      <label
-        htmlFor={name}
-        className="block text-[10px]"
-      >
-        {label}
-      </label>
-      <Controller
-        control={control}
-        name={name}
-        defaultValue={date?.format('YYYY-MM-DDTHH:mm')}
-        render={({ field }) => (
-          <input
-            type="datetime-local"
-            {...field}
-            min={now}
-            className="block w-full border-0 p-0 focus:ring-0"
-            onChange={(e) => {
-              // Convert the local datetime to UTC
-              const date = moment.utc(e.target.value);
-              if (setDate) setDate(date);
-              field.onChange(e.target.value);
-            }}
-          />
-        )}
-      />
-    </div>
-  );
-};
